member expressions in predicates were missing from the type table
Summary:
calling `expression cx expr` determines the type of `expr` and adds it to the type table. due to sentinel prop checking, we never call `expression` on member expressions in predicates like `if (foo.bar)` or `foo.bar == 'a' ? 1 : 2`, because we need the type for `foo` and don't want to walk things twice.

we were neglecting to manually add `foo.bar` to the type table, which meant that type-at-pos, dump-types and coverage didn't know about those expressions.

Reviewed By: samwgoldman

Differential Revision: D6390968

fbshipit-source-id: 16fd31bcfb036f21e377eb0a8a44e0eb87771830