[hack] Add a few ImmQueue operators
Summary:
I want the 'concat' operator for ImmQueue.

(reason: I have built up a ImmQueue of outstanding LSP messages that have yet to be sent to the flow server. When I finally succeed in starting the flow server and connecting to it, I must send two messages followed by the remaining ImmQueue. This is why I want concat).

Reviewed By: arxanas

Differential Revision: D6438987

fbshipit-source-id: c6f4a5d21947770c730e79d16ffb64336451a684