Server reports exit status over persistent connection
Summary:
We want the Flow LSP client to automatically restart the flow server in case it shut down due to a change in flowconfig. But in all other cases where the flow server shuts down, the LSP client will merely tell the user "the flow server has shut down" and give them a convenient Restart button.

(Motive: shutdowns might happen for so many reasons\! if the user manually did "pkill flow" at the command line then we don't want to restart it; if flow is crashing then we don't want to get into a restart cycle. So we'll think about this as a WHITELIST: we'll only do an automatic restart if the shutdown was a clean shutdown due to one of the auto-restartable conditions).

In unix the client is unable to obtain the exit status of the process. So, instead, we'll have the server broadcast its exit status to all persistent connections. It won't do that in all exit reasons, but it will at least do it for the exit reasons on the whitelist.

Note: I believe that if we write to stdout and then do a clean exit, then standard library behavior is to flush all output buffers before completing the exit. I'll have to test this.

Reviewed By: nmote

Differential Revision: D5798983

fbshipit-source-id: 385e713401449768a90dad1518e044b4a474bbee