Refactor start_flow_server so it doesn't exit
Summary:
Previously, the method "start_flow_server" returned (1) true if it managed to start the flow server, or (2) false if someone else had started it first, or (3) exit the program in case of other non-recoverable non-diagnosed error case.

I've changed it to raise an exception in case (3), and made its (only) caller be the one to exit the program in case of exception.

This is so I can call into start_flow_server from the LSP server, which should report an error rather than exiting.

Reviewed By: mroch

Differential Revision: D5801880

fbshipit-source-id: 285720bb0a5492b94f2df25a52f88de9fe8ffdb5