Fix recheck bug in lazy mode
Summary:
In lazy mode we ignore dependents of dependencies. These files
are not checked, unless we care about them for some other reason.

The bug was that "some other reason" should be "the file is focused, a
dependent, or a dependency" but was "the file is focused or a
dependent".

This manifested during rechecks in files which were dependencies of
files we care about AND dependents of other dependencies. During
recheck, we were failing to recheck those files, leading to weird errors
when class IDs no longer matched.

Reviewed By: samwgoldman

Differential Revision: D6490385

fbshipit-source-id: 0014584a95cca21b9fd87d99ddf2e1eeda15ce03