Do not abort find-refs when encountering untyped code
Summary:
If, in the code `foo.bar()`, `foo` is untyped, and we are searching for a member
`bar`, the entire find-refs request would previously fail. Now, it will just
ignore such potential matches.

In the future it might be nice to report these in the UI as potential matches,
but for now we'll just ignore them since there's no API or UI to surface them.

Of course, this only has an effect where the file is checked but some code
within it is poorly-typed. We don't look at unchecked files at all.

Reviewed By: samwgoldman

Differential Revision: D6482281

fbshipit-source-id: a3791f12460bc55d1dfe656d4a030f874fe5cfb6