[parser] microoptimize Peek apis
Summary:
optional params are heap allocated. the `Peek` apis take an optional `i`, `?(i=0)`, which they thread through to `token`. for each one, it unwraps the arg and allocates a new option for the nested call.

I don't expect this to make any difference, but I did notice the allocations looking at heap profiles of flow_parser.js in V8 and this seems not-so-bad.

Reviewed By: samwgoldman

Differential Revision: D6454158

fbshipit-source-id: ec81b1afd9a04df32532c584078e3b56cc4511f6