Fix race condition in computed property initialization
Summary:
Computed properties in object initializers are processed like a set-prop. If
the key type is a tvar, the write is defererd until the tvar resolves to
something.

It's important that these writes happen before any reads, otherwise we might get
spurious "property not found" errors.

To induce an order, this diff uses the same trick as spreads. That is, we create
a tvar which is the output of the object-assign operation and holds later
operations at bay until the object-assign happens.

Similarly, we create a tvar which is the output of the set-prop operation, so
future operations are held at bay until it happens.

Note that set-prop does not have an "output" position, so this diff adds an
optional one. The only user of this feature is object literal initialization.

Note that this is a pretty surgical fix, which leaves existing bad behavior in
place. For example, a computed property with a union-of-string-literal type
should create a union of object types, but today we create an object type with
both properties.

Fixing related issues is left for future work. Such fixes should make the
optional output tvar in set-prop unnecessary.

Reviewed By: avikchaudhuri

Differential Revision: D6425905

fbshipit-source-id: 52987f3e59196d7b07b2a15f720ea550281b2842