Remove `export ... from` module refs from module sig require lists
Summary:
Before this diff, when we saw an `export ... from` statement in an ES module,
we added the module ref to the list of ES imports in the file sig requires map.

This change removes the entries from the ES imports list, which means we need
to also walk exports to find all file dependencies. The `require_loc_map`
helpers has been extended to perform this walk, and a number of call sites have
been changed to use this helper method instead of walking only requires.

This change, combined with a follow-up change to make requires preserve even
more information will be used when calculating file dependencies and exported
names.

Reviewed By: nmote

Differential Revision: D6412073

fbshipit-source-id: 8000503e47b667953d27a71a2d8718212f9e39a8