flow cycle subcommand
Summary:
Dependency cycles can be bad for Flow performance. This diff introduces a new
subcommand that makes it easier to extract useful information from Flow's
dependency information.

In particular, users can point to any file in a cycle and the command will
output a DOT file to stdout. This file can be fed into a program like Gephi to
to further analysis.

Reviewed By: gabelevi

Differential Revision: D6202923

fbshipit-source-id: 94f6c7862443f0cf0ee8fbfdd22e79504e584dda