Better error message when file is required as input
Summary:
In several commands a .js file is required as input (coverage, dump-types
commands, etc.), instead of a file or folder. Until now the error message
returned when a directory was passed in as argument was
```
Sys_error("Value too large for defined data type")
```
This diff fixes this so that a more appropriate error message is returned.

Reviewed By: samwgoldman

Differential Revision: D6504960

fbshipit-source-id: 1e4f5b5dd4d866663bbd403984362d713424b55a