[hack] functorize Hh_json printing, make escape add directly to buffer/oc
Summary:
`Hh_json` is able to write to a buffer or an out_channel, but the code was duplicated. This diff introduces a functor and deduplicates most of it (except multiline, which has different behavior between writing to a string and oc).

Also, `escape` previously allocated a new buffer, converted it to a string, then either wrote it to a buffer or an out_channel. Now it writes directly to the buffer or oc.

Reviewed By: samwgoldman

Differential Revision: D6531053

fbshipit-source-id: dc5fdafa94ea8e4e98a08856641a882058ad526a