Drop constructor_annots
Summary:
This subtyping rule was added to support a use case that is now uncommon:
"casting" a ES5-style class (function) to a ES6 class.

This rule depends on interfaces supporting static properties (specifically,
static callable properties), but in order to simplify and solidify the rules
around interfaces types we should drop the support for statics, such as it is.

Reviewed By: avikchaudhuri

Differential Revision: D6528007

fbshipit-source-id: 1a7d473f9b866ce52e3a661a6d63c94049860e18