Improve error message for unexpected strict modifier
Summary:
Instead of showing an "unexpected token" error on the next token, we can be
more specific about what is wrong by pointing to the `static` keyword itself.

This error message improvement is particularly useful, since a diff down-stack
makes statics on interfaces a parse error (instead of a poorly implemented
feature).

Reviewed By: mroch

Differential Revision: D6552427

fbshipit-source-id: 9442befb96308894cec38eed19b2175af3110731