Make type utils part of TypeTerm recursive module set
Summary:
It's not currently possible to use utilities like reason_of_t from within the
recursive modules in type.ml, which means we sometimes need to jump through
hoops in order to write the code we want to write.

This diff simply moves a few useful functions from the bottom of type.ml into a
module TypeUtil, which part of the mutually recursive set of modules in
type.ml.

Reviewed By: gabelevi

Differential Revision: D6482245

fbshipit-source-id: d9860ddf9bfa720f821b22cac7214cc61a8adae5