Collect evaluated tvars before union speculation
Summary:
When a union appears in an upper bound position, we walk types attempting to
concretize any tvars. We need to concretize these tvars in order to perform
speculation.

Before this diff, we never walked any EvalT type. This diff adds a subset of
EvalT tvars which we want to treat as 0->1. Specifically, type destructors,
which arise from annotations, will eventually resolve to a type.

Note that this is a partial fix. For one, we might not have created the tvar
yet. Also, other EvalT types (DestructuringT) should also be walked, but since
they are not 0->1, we would need to treat them as unresolved.

Further note: Not all type destructors are actually 0->1, because type
parameters are not 0->1. The existing collect_of_type code assumes type params
are 0->1, so I'm extending that shortcoming here as well.

Reviewed By: gabelevi

Differential Revision: D6456035

fbshipit-source-id: 0edf3a1e4c2d4bef29e88b31942cbd20495de3f3