Add location to shadow prop reasons
Summary:
Shadow properties are added when an unsealed object reaches a read or a write
for an unknown property. While it is safe to use a locationless reason here,
because these reasons will never be part of an error message, it's useful for
debugging in verbose mode to know why the property was added.

This diff uses the "op" reason -- corresponding to a get-prop, set-prop, or
subtyping constraints -- to provide a location for the reason.

Reviewed By: gabelevi

Differential Revision: D6524317

fbshipit-source-id: 4b6a6d6903d23d76ba4c83d89d9da11a1ae31046