Lib files and directories should be implicitly included
Summary:
At the moment, lib directories and files aren't watched by
Flow's file watcher, but includes are. So if a lib is under the root
directory or under an included directory, Flow will notice when it
changes. But if you don't include an external lib directory, Flow won't
notice when those libs change.

This changes Flow to treat libs as being implicitly included.

Reviewed By: samwgoldman

Differential Revision: D6550439

fbshipit-source-id: 822d279c674657c6547a257605735c4a6419d129