[PR] [experimental] flow-parser-bin node extension using native parser
Summary:
The js_of_ocaml version of the parser, `flow_parser.js`, is not as efficient as the native ocaml version.

OCaml and node can both interface with C/C++ code, so we can expose the parser as a C++ library (see changes in src/parser/) and then consume it from a node extension (see packages/flow-parser-bin/).
Closes https://github.com/facebook/flow/pull/5508

Reviewed By: gabelevi

Differential Revision: D6552926

Pulled By: mroch

fbshipit-source-id: 5b491cb8d88fcbc2517b5e891589bcfe082ac74a