Add loc of default keyword to default exports in file sig
Summary:
Currently, default exports don't always have a clear location to point to. For
example, given `export default <expr>` it's not necessarily a good idea to
point to `<expr>`, which could span vast swaths of source.

Consider get-def, which might resolve to a default export. This location gives
a sensible place to jump.

This location is also useful for follow-up work that refactors import/export
code. We can pre-populate tvars based on file sig information, but we need a
location so the statement pass can look up the tvar later. Default exports are
the only export type that don't always have a location without this change.

Reviewed By: nmote

Differential Revision: D6541404

fbshipit-source-id: 3df344232de06c1632598e90417130fb485f2e5e