Stop collating errors during init and recheck
Summary:
Back in the day, we used to collate errors before sending them to a
client. But this made commands like `flow status` slower than they
needed to be. Furthermore, the Flow server would be busy collating
errors without a .init or .recheck lock, and this would cause trouble.

So we started caching collated errors and always regenerated them during
init or recheck.

Now that we have the monitor, we no longer need to regenerate them
during init and recheck. We can make it lazy and only regenerate them
when we actually need to.

Reviewed By: samwgoldman

Differential Revision: D6572696

fbshipit-source-id: 65f042820e4dd47f4309bf5098363b1ef2981d8c