Property type probe in MethodT and SetPropT
Summary:
The method identifier at method calls is currently not being tracked
in the type tables. However, it would be useful to get type info about
the method in such cases. Same holds for the field being updated at
set property operations.

The way constraint generation worked until know, there was no way
to access this type. This only became available after the respective
flows were handled by constraint propagation.

To fix this we introduce a probe type as an optional last element of
`MethodT` and `SetPropT`. This type is set up at constraint generation,
(`statement.ml`) and is populated appropriately when the respective
constraints are evaluated (`flow_js.ml`).

This machinery should not affect the rest of the constraint graph, so
ought to have zero diff on error reports.

Reviewed By: samwgoldman

Differential Revision: D6586944

fbshipit-source-id: d14c14e44d5ae755c867a4d9ed5b372cb9c46828