Only call merge_contents_context rather than basic_check_contents
Summary:
This paves the way to parallelize find-refs. We can't effectively call
basic_check_contents in parallel because it too requires workers which cannot be
serialized as part of a job (and would probably break in subtle ways even if
they could be). merge_contents_context gets the results we need with less
computation and no workers.

Reviewed By: gabelevi

Differential Revision: D6574065

fbshipit-source-id: 3a7f16c411da0961710ccc77efac3aa2efd11b71