Deprecate legacy `declare var exports` syntax
Summary:
In Flow v0.25, we added a new and improved syntax to describe the exports
object of a declared CommonJS module. Instead of `declare var exports: T`,
users should write `declare module.exports: T`.

The new syntax makes it possible to declare a CommonJS module that exports an
object with an `exports` property.

However, we never officially deprecated the old syntax, so now we have both,
which is confusing.

In order to provide a gentle transition, this diff makes the old syntax a lint
error (by default, no config needed). Let's roll a couple versions with this
lint error before officially dropping support.

Because this is implemented as a lint error, it can be disabled project-wide in
the `[lints]` section of a .flowconfig. While I hope that people will simply
fix the errors, it's possible that the deprecated syntax comes from a
dependency which needs to be updated first.

Reviewed By: gabelevi

Differential Revision: D6580299

fbshipit-source-id: 06d7f4fe34d3ffb2c2a32b1d8c446c5f67b82a49