Kill merge job after timeout
Summary:
If a component takes more than 100s to infer, we should complain loudly.

This diff piggy-backs on the timer that logs slow merges, which fires every
15s. After 100s (in practice, 105s), we give up and raise an exception, which
kills the worker and adds an internal error to the server env.

Because we can recover gracefully from merge job exceptions, the rest of type
inference proceeds as usual, with the aborted contexts treated as `any`.

Reviewed By: gabelevi

Differential Revision: D6605673

fbshipit-source-id: ef83b2fd0e874e74c1eb0b80e489223cbbd97f74