Remove "some incompatible instantiation" language
Summary:
The reason description "some incompatible instantiation of `T`" is difficult to parse. It makes more sense to see "`T`" alone. I've been waiting to make this change since the "some incompatible instantiation" message is a great sign-post for the people working on Flow that the error is created from the `empty` when we run tests for polymorphic types.

For what it's worth, if we compare polymorphic type parameters to opaque types as panagosg7 proposes then this reason description is consistent.

Reviewed By: gabelevi

Differential Revision: D6516208

fbshipit-source-id: ab5ad35a5665bef45f4290ee4196fa27c6bd6bfe