Re-construct use_ops in actions after speculation
Summary: During speculation we replace `use_op` with `UnknownUse`. We do some checking, build a new `use_op` chain, defer some actions, but when we go to fire those deferred actions we don't replace the `UnknownUse` we set with the original `use_op` that got replaced. This diff fixes that.

Reviewed By: gabelevi

Differential Revision: D6548361

fbshipit-source-id: fa50752f87e29da3aafc957aea907f8273e28cc1