Add use_op frame before React config checking to increment score
Summary:
In practice, there are a lot of React components generated by HOCs that are an intersection of a component type and some other object or interface type which adds static or instance properties. e.g.

```
import * as React from 'react';

declare var MyComponent: React.ComponentType<{foo: number}> & {
  someStatic: boolean,
};

<MyComponent />;
```

This diff adds a `use_op` when we enter React config checking so that those errors win when we pick the errors to show a user after speculation.

Reviewed By: gabelevi

Differential Revision: D6448816

fbshipit-source-id: e9986f180be6d759ad16eb104a5b529ac217bd5e