make get-def see through requires
Summary:
The get-def command can already "see through" a local definition corresponding to
an ES6 import to the corresponding export in the imported module.

But we didn't have the same behavior for local definitions corresponding to
requires, since the bindings are of the normal kinds (var / let / const) and the
general behavior for such kinds of bindings is to *not* see through their
initializers.

This diff special-cases requires to match the behavior of ES6 imports:

(1) We provide get-def with the ability to return an intermediate result (as an
alternative to a final result) which is then chained into a subsequent get-def.

(2) We recognize local definitions corresponding to requires, and whenever such
a definition would be previously returned by get-def as a final result, it is
now returned as an intermediate result.

Another approach that was considered was to add a bit to the binding when it was
created, somewhat like ES6 imports have, but it was much more messy to carry
around that bit just for this special case.

Reviewed By: samwgoldman

Differential Revision: D6625840

fbshipit-source-id: 8e9e1348fcc3b842fd08b5c41d74687bf6b92606