stripped down topsort
Summary:
Ordering components by "height" (which approximates the length of the critical path of a component
in the dependency graph) is redundant in the current dynamic scheduling algorithm for the type
checker. (It might have been used at some point in a static scheduling algorithm that ran components
at the same height in parallel, after running components of smaller heights.)

The true topological order, if needed in the future, can be recovered from the existing parts of the
state of the topsort algorithm.

Reviewed By: samwgoldman

Differential Revision: D6652191

fbshipit-source-id: 7ac8e342a8b4d5bda3e4cc11df6b070825574aba