Segregating bookkeeping of coverage from type queries
Summary:
The goal here is to provide more fine grained type information for
`type-at-pos`. To get there we need to segregate bookkeeping of
location to type bindings for the purpose of coverage results from
that of type queries.

To do so we introduce a second table in `type_table.ml` that is geared
towards `type-at-pos`. The locations that enter this table are expected
to be identifiers (this includes symbols that might appear in type
positions and some keywords, e.g. `this`, `super`), instead of any
possible expression.

Reviewed By: samwgoldman

Differential Revision: D6554319

fbshipit-source-id: 633e8a6db6d67a3cfc1d352c23c05801cd0865e0