Fix deprecated-default-exports lint rule default setting
Summary:
Before this change, in D6580299, I changed `LintSettings.default_severities` to
include the `deprecated-default-exports` lint as an error.

However, the default settings value was also used when creating the
`.flowconfig` file in the `flow init` subcommand. If the `.flowconfig` was
created this way, the `[lints]` section would have an explicit lint turning on
the `deprecated-default-exports` lint, which would subsequently error because
specifying a lint that is in `default_severities` is superfluous.

This change renames `default_severities` to `empty_severities`, reflecting how
that value is actually used. The `deprecated-default-exports` lint is defaulted
to on in the base flowconfig.

Reviewed By: gabelevi

Differential Revision: D6643364

fbshipit-source-id: f4ceb7dfa8d7595602abe1f0c4c76d54b1b29121