Type components as a non-empty list
Summary:
We pass around a list of file keys, which we expect to be non-empty everywhere.
We even call `List.hd` on this list, which raises for the empty list.

Instead of relying on diligence to preserve this invariant, we can use the
`Nel.t` type to let the OCaml type checker do the (not so) heavy lifting.

Reviewed By: panagosg7

Differential Revision: D6678139

fbshipit-source-id: d316b6859f64ab55a05a3954f7fcb826158c6c83