prevent appveyor from overwriting github release description
Summary:
in D5041001, we added "force_update: true" to the appveyor deploy config in an attempt to make it upload the Windows artifacts to an existing release, which didn't seem to be working at the time.

this caused appveyor to remove the existing description (see https://github.com/appveyor/ci/issues/2017, thanks IlyaFinkelshteyn!)

it sounds like appveyor should warn about an existing release but still upload artifacts, if we don't use `force_update`.

Reviewed By: nmote

Differential Revision: D6686134

fbshipit-source-id: d0c744d7bdf8cde001bea191088758df7c67323f