[hack] Add new flag to Hack to ignore the build revision when loading saved-states.
Summary: In order to test changes to Hack saved-states, it's necessary to base the changes to be tested on top of a diff such as differential/diff/39743908/ that will enable saved-state loading, since normally the build revision for local Hack builds is empty.  Instead, the process should be formalized and checked in as an option that all Hack developers can use easily.

Reviewed By: dabek

Differential Revision: D6659049

fbshipit-source-id: fb09f9a09ab61f8602d776119908cced58ae645f