Add --merge-timeout server flag
Summary:
By default, Flow will throw an internal error if a single file
or cycle of files takes over 100s to merge. This makes that timeout
configurable.

I decided to leave it out of the .flowconfig for now, since I don't need
that functionality and I'm not sure whether I like the idea of a whole
repo changing it's timeout. This diff is mainly to allow tooling to
relax the timeout.

Reviewed By: samwgoldman

Differential Revision: D6708346

fbshipit-source-id: 2dfa516b2c3667f791fdc6222c15c17ccd959735