Better lwt errors for select's EINVAL
Summary:
This is tricky. There's a few complicating issues:

1. OCaml will raise EINVAL in a nonstandard manner. Reported it here: https://caml.inria.fr/mantis/view.php?id=7700
2. Lwt handles EBADF but not EINVAL. Reported it here: https://github.com/ocsigen/lwt/issues/529
3. The `select()` systemcall doesn't work with fds >= 1024

The third point is the tricky one. We can repro it like so:

1. Start up a flow server on some large repo
2. `for i in `seq 2000`; do flowd status --quiet ~/repo > /dev/null & done`

This will create more than 1024 clients connected to the monitor, all
waiting for the server to finish initializing. At some point, one of
their file descriptors (which are ints) will be >= 1024 (FD_SETSIZE).

So ultimately we have two choices:

1. Continue using `select()` and enforce that we never have more than 1024 fds open in the monitor
2. Add libev as a system dependency and use that instead of `select ()`. And figure out what to do for Windows

Reviewed By: samwgoldman

Differential Revision: D6639011

fbshipit-source-id: d9e5f3c560b5a629971cc3d683da813904323ab3