Limit concurrent client connections to the monitor
Summary:
The Flow server monitor uses `select()`, which can't deal with
fds >= `FD_SETSIZE`, which seems to be 1024 everywhere unix-y (not sure
what it is on Windows. You can currently cause Flow to crash by
connecting > 1024 clients to the server at once.

The easiest fix is to limit the number of connections that we accept. We
could do this by counting connections, but a slightly easier way is to
reject any fd > 500. If we later introduce fds in some other way, this
should help keep the total fds constant.

500 is completely arbitrary.

Ideally, we'd be able to switch to `poll ()` via libev or something like
that, but for now this is better than crashing the server.

Reviewed By: samwgoldman

Differential Revision: D6639366

fbshipit-source-id: c388e7c6900d3cecaa24ef7b553d405aeb70af8f