Reintroduce abstract_targ
Summary:
In D4620957, I removed this code because it seemed unused. Indeed, at the time
we had no code that hit this logic and there was no test that exercised it.

While looking into supporting a `Thenable` or `PromiseLike` type, I ran into
non-termination which I reduced to a test case, included. This pattern of
structural checking on type apps is not detected by the typeapp expansion cache
without the abstract_targ logic.

Reviewed By: gabelevi

Differential Revision: D6645629

fbshipit-source-id: 08f841912721703f1c27fea4db48788393ab2e32