Support implements for declared classes
Summary:
We have supported `implements` on class declarations for a while now. The
implements logic has two nice properties:

1. Subtyping can be checked at declaration time instead of use time. This is
important for library definitions which might not have any uses. It's also just
nice to be able to find and fix errors while working on the class declaration
instead of later on, when using the class.

2. Structural checks are potentially expensive, but a class declared to
implement some interface only performs a structural check once -- at the
declaration site. From then on, we can compare the nominal ID from the
implements list to the nominal ID of an interface.

This diff simply extends the capability that has existed for a while to
declared classes. I plan to use this capability to define a `PromiseLike`
interface, which `Promise` can explicitly implement, in order to support custom
promises with good performance characteristics.

Reviewed By: gabelevi

Differential Revision: D6645705

fbshipit-source-id: 952e19d0eb3ed114606a82cb65913213cfa9e28f