Calculate unchanged in Merge_stream.join
Summary:
Before this change, a list of unchanged files was accumulated in the merge job,
in order to revive the merge batch in the merge worker's merge/join procedure.

In a follow-up diff, I want to make the sig context and sig hash heaps more
granular, one for each file instead of one for each component. This would
require passing more information back to the master process in the form of
unchanged files -- because part of a cycle can be unchanged.

Instead, let the master re-calculate what has changed based on the old and new
sig hashes, which are added by the merge workers. This does involve extra work
in the master process, but this is not expensive in practice.

Note that we still need to accumulate a list of skipped files, because those do
not appear in the merged list but still need to be revivied.

Reviewed By: gabelevi

Differential Revision: D6687150

fbshipit-source-id: bbded86849af5da1ab3883c3eefd21bd1787a900