[PR] Add Code of Conduct
Summary:
In the past Facebook didn't promote including a Code of Conduct when creating new projects, and many projects skipped this important document.

The Flow docs website links to the Code of Conduct in the website footer, which is great! :)

It can be even more clear to contributors if also mentioned in the `CONTRIBUTING.md` and in a `CODE_OF_CONDUCT.md`, as per the Github community guidelines.

**why make this change?:**
Facebook Open Source provides a Code of Conduct statement for all
projects to follow, to promote a welcoming and safe open source community.

Exposing the COC via a separate markdown file is a standard being
promoted by Github via the Community Profile in order to meet their Open
Source Guide's recommended community standards.

As you can see, adding this file will improve [the Flow community profile](https://github.com/facebook/flow/community)
checklist and increase the visibility of our COC.

**test plan:**
Viewing it on my branch -
<img width="1497" alt="screen shot 2018-01-14 at 3 33 30 pm" src="https://user-images.githubusercontent.com/1114467/34922133-9e0942f2-f940-11e7-938d-2e8cb5ca2748.png">
<img width="1510" alt="screen shot 2018-01-14 at 3 33 51 pm" src="https://user-images.githubusercontent.com/1114467/34922134-9e288248-f940-11e7-9193-2d9f88b08d1c.png">

**issue:**
internal task t23481323
Closes https://github.com/facebook/flow/pull/5660

Reviewed By: samwgoldman

Differential Revision: D6728591

Pulled By: flarnie

fbshipit-source-id: 4838f6bcf0da75957b083974c78710187ee6047a