Add Unused_server exit code
Summary:
Once a day, we check if the Flow server hasn't had a client
connect in 7 days. If so, we exit with exit code 5. Since this is a
non-standard exit code, this confused the Flow monitor and appeared in
the logs as an Unknown_error.

The easy fix is to just standardize this exit code.

Reviewed By: samwgoldman

Differential Revision: D6638520

fbshipit-source-id: 78c7bf22414f050205b922e9d5ae8e3d2d26d06c