Fix recheck opts in lazy mode
Summary:
`Types_js.typecheck` receives an argument `infer_input`, which is the sets of
focused, dependent, and dependency files which need to be rechecked. Due to
laziness, the `infer_input` sets might include files whose dependencies we have
not yet checked. Furthermore, the dependents of files in `infer_input` might
have unchecked dependencies.

The logic to handle this was added for lazy mode (nee quick start mode) by Avik
in D5067264. This logic worked by first adding `all_dependent_files` to
`infer_input`, then adding all the dependencies of those files.

Later, in D5863553, Gabe made the clever observation that laziness can help us
avoid checking the potentially many files in `node_modules`, unless they are
dependencies. This made the lazy logic apply to non-lazy mode as well.

However, there was a problem with the original implementation in lazy mode,
and, after D5863553, in non-lazy mode as well. Adding `all_dependent_files` to
`infer_input` forces the `recheck_map` to include a `true` entry for all files,
which prevents recheck optimizations from happening.

The `recheck_map` is created in `make_merge_input`, which is parameterized by
`infer_input`. For rechecks, that function creates a `to_merge` set by adding
`all_dependent_files` (making it unnecessary to add in `typecheck`), but also
creates `recheck_map`, which returns `true` for any file in `infer_input` or
the `direct_dependent_files`, which is a subset of `all_dependent_files`.

The reason that `recheck_map` exists is to skip rechecks of components whose
dependents have not changed. We track whether we need to recheck based on the
`recheck_map`, so if it returns true for every file, we never skip a component.

This diff makes a very simple change. In `typecheck`, instead of adding
`all_dependent_files` to `infer_input`, only add the dependencies that we
calculate based on `infer_input + all_dependent_files`. Remember that
`make_merge_input` will add `all_dependent_files` back in anyway.

This fix exposed a couple other issues downstream. The biggest issue was that
skipped components did not unblock blocked components from the merge stream.
The solution is trivial -- just use the existing `push` helper.

The second issue is more subtle. The flow server monitor tracks how many files
have been merged, but treated skipped modules as `0`. However, the calculation
of total files contains all components, because we don't know if they will be
skipped or not. To make the counts accurate again, I simply added the size of
the skipped component to the accumulator.

Reviewed By: gabelevi

Differential Revision: D6742651

fbshipit-source-id: 56e56cb3b65342abbd3928fa893c8b30698c0aa1