Remove unnecessary accumulator for unchanged files in merge job
Summary:
We only need to return the unchanged file list to the `merge/join` function, which
uses the information to revive oldified heap entries. It's unnecessary to also
accumulate and return the unchanged file list from the worker. Indeed, the
merge service throws away the second element of the accumulated tuple.

In D4302807, Marshall changed the signature of MultiWorker, which enabled a
worker to accumulate a value of a different type to the job's return type. I
assume that, because this code predates that change, this accumulation was
simply to satisfy the type checker.

Reviewed By: mroch

Differential Revision: D6686849

fbshipit-source-id: a28848ceb47fb1592415287fb93ac4479c961ad2