Fix sig hashing of reasons
Summary:
`Marshal.to_string` takes two arguments, but we were only passing one. This
meant we were actually passing a function pointer to Xx.add. Every OCaml value
supports the String_val macro, so this error was not detected.

Unlike the add_int API, which is unsafe on purpose to support hashing values
which are ints internally (bool, constant constructors), the add_string API is
always called with values that are actually typed as strings.

Reviewed By: gabelevi

Differential Revision: D6762516

fbshipit-source-id: 225cbc3f1b48c76d82be19bcd9800029c47daf1c