[hack] Remove Watchman sync directory
Summary:
Before the Watchman flush-subscriptions command, we simulated synchronous
queries by touching a file on disk and waiting for that file notification
to show up on the Watchman subscription. That sentinel file we placed
inside the sync_directory (which usually is the .hg directory).

We no longer have a need for it now that flush-subscriptions has been in
use for quite some time.

As an aside, we can also stop watching the .hg directory.

Note: we still end up typechecking files inside the .hg directory in "--check"
mode since Hack doesn't use Watchman in check_mode

Reviewed By: dabek

Differential Revision: D6708987

fbshipit-source-id: 336ed5ebbcf122e00246e2d69b3217a4219e9270