Fix nonstrict import rule when in a cycle
Summary: The nonstrict import rule occasionally does not work correctly in cycles. This is because when we are in a cycle, during the merge we use the context of the "leader" module, and thus the is_strict flag won't necessarily be accurate. On samwgoldman's suggestion, I have added a is_strict flag to the import/require types, and then we use that rather than the context when checking the nonstrict import rule.

Reviewed By: samwgoldman

Differential Revision: D6788215

fbshipit-source-id: b116e5a9d32d2b2fc67d33be0ca59be30a0f94e5