Remove unused "should_merge" logic
Summary:
We've now combined the infer and merge phases of the typechecker. Before that
change, when lib defs errored, we would perform local inference, collecting as
many errors as we could, but not merge, because the libs were broken.

Now, after combining infer and merge phases, it is no longer possible to
perform local inference and not merge. This change simply removes the plumbing
that supported that use case.

Reviewed By: gabelevi

Differential Revision: D6828067

fbshipit-source-id: a76b90e204ca5725d4119e62ae02a652481b8873