[hack] Add names to Local_id.t
Summary:
Currently, we have a flag to enable name tracking for local identifiers because of concerns about performance when this tracking is enabled.

I don't think it's necessary to track names in a map--we can simply store names inside the Local_id.t type. This way, consumers of the NAST or TAST who would like access to the names of local variables can look them up without needing to rely on hidden global state (which is not always enabled).

Reviewed By: dlreeves

Differential Revision: D6756078

fbshipit-source-id: 1c3d9556f85dc6a6a6fdf7d81a48fdbbef80cc4a